Lidex_Api/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.8.0
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
Pillow==9.4.0
platformdirs==2.5.2
psycopg2==2.9.5
PyAutoGUI==0.9.53
pycodestyle==2.10.0
PyGetWindow==0.0.9
pymongo==4.3.3
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.14
virtualenv==20.16.5
Werkzeug==2.2.3
whitenoise==6.2.0
wikipedia==1.4.0
